inventory: Fixes #1913 Don't send inventory from PXE
Actually, last inventory is always displayed on web interface.

If an inventory already exists in DB for a machine identified
by its MAC adress, don't set this inventory as the last one
if it is a PXE inventory.

So, PXE inventories will not be displayed anymore as
the "last"
inventory.