scheduler: Add imaging option
Fix related to ticket #2075 broke package deployment

When deploying, if wake-on-lan is done, connection to mmc-agent is made.
So, we have to know if computer is linked to an imaging server or not.

We have to think about it...

So, quick fix is to add this imaging option (disabled by default) who
don't call mmc-agent.

Enable it to enable feature #2075