msc: Refs #2255 Fix phases creation in convergence
When a machine is added to a convergence command, and if this machine
was already deployed, phase creation failed.

One machine can be part of the same command several times with
convergence, so create phases only for machine who will be added,
and not reset all phases for this machine.